# backend/template/src/api/services/models/office.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Office {
  created_at?: string;
  lastModifiedBy?: number;
  location?: string;
  name?: string;
  officeId?: number;
  phoneNumber?: string;
  updated_at?: string;
}