# backend/template/src/api/services/models/organization.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Organization {
  address?: string;
  averageRate?: number;
  created_at?: string;
  email?: string;
  image?: string;
  is_deleted?: boolean;
  lastModifiedBy?: number;
  name?: string;
  organizationId?: number;
  phone_number?: string;
  updated_at?: string;
}