# backend/template/src/api/services/models/services.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Services {
  created_at?: string;
  description?: string;
  lastModifiedBy?: number;
  serviceId?: number;
  serviceName?: string;
  type?: string;
  updated_at?: string;
}