# backend/template/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"Cloud nine - queute - ESPRIT PI 4ArcTic","description":"OpenApi documentation for Queute App - PI ESPRIT 2024-2025","contact":{"name":"Sabta Ruqaia","email":"ruqaia.sabta@esprit.tn"},"version":"1.0"},"servers":[{"url":"http://localhost:8082/api/v1","description":"Local ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/feedbacks/{feedbackId}/mark-as-read":{"put":{"tags":["feedback-controller"],"operationId":"markFeedbackAsRead","parameters":[{"name":"feedbackId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Feedback"}}}}}}},"/services/create/office/{officeId}":{"post":{"tags":["service-controller"],"operationId":"createService","parameters":[{"name":"officeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Services"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/organisations":{"get":{"tags":["organisation-controller"],"operationId":"getAllOrganisations","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Organization"}}}}}}},"post":{"tags":["organisation-controller"],"operationId":"createOrganisation","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"requestDTO":{"$ref":"#/components/schemas/Organization"},"imageFile":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Organization"}}}}}}},"/offices/assign-to-organisation/{organisationId}":{"post":{"tags":["office-controller"],"operationId":"addOfficeAndAssignToOrganisation","parameters":[{"name":"organisationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Office"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Office"}}}}}}},"/feedbacks/organization/{organisationId}":{"get":{"tags":["feedback-controller"],"operationId":"getFeedbacksByOrganization","parameters":[{"name":"organisationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Feedback"}}}}}}},"post":{"tags":["feedback-controller"],"operationId":"addFeedbackToOrganization","parameters":[{"name":"organisationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Feedback"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Feedback"}}}}}}},"/feedbacks/feedbacks/respond/{feedbackId}":{"post":{"tags":["feedback-controller"],"operationId":"respondToFeedback","parameters":[{"name":"feedbackId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/services/update/{id}":{"patch":{"tags":["service-controller"],"operationId":"updateService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Services"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Services"}}}}}}},"/organisations/{id}":{"get":{"tags":["organisation-controller"],"operationId":"getOrganisationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Organization"}}}}}},"patch":{"tags":["organisation-controller"],"operationId":"updateOrganisation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"organizationDetails":{"$ref":"#/components/schemas/Organization"},"imageFile":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Organization"}}}}}}},"/offices/updateOffice/{id}":{"patch":{"tags":["office-controller"],"operationId":"updateOffice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Office"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Office"}}}}}}},"/services/get/{id}":{"get":{"tags":["service-controller"],"operationId":"getServiceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Services"}}}}}}},"/services/all":{"get":{"tags":["service-controller"],"operationId":"getAllServices","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Services"}}}}}}}},"/organisations/{organizationId}/average-rate":{"get":{"tags":["organisation-controller"],"operationId":"getAverageRate","parameters":[{"name":"organizationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"number","format":"double"}}}}}}},"/organisations/offices-count":{"get":{"tags":["organisation-controller"],"operationId":"getOfficesCountByOrganization","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}}}}}},"/organisations/images/{filename}":{"get":{"tags":["organisation-controller"],"operationId":"getImage","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string","format":"binary"}}}}}}},"/organisations/global":{"get":{"tags":["organisation-controller"],"operationId":"getGlobalStatistics","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/organisations/counts":{"get":{"tags":["organisation-controller"],"operationId":"getAllOrganizationsEntityCounts","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationEntityCountsDTO"}}}}}}}},"/offices/{id}":{"get":{"tags":["office-controller"],"operationId":"getOfficeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Office"}}}}}}},"/offices/by-organisation/{organisationId}":{"get":{"tags":["office-controller"],"operationId":"getAllOfficesByOrganisationId","parameters":[{"name":"organisationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Office"}}}}}}}},"/feedbacks/user/feedbacks":{"get":{"tags":["feedback-controller"],"operationId":"getFeedbacks","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeedBackDTO"}}}}}}}},"/services/delete/{id}":{"delete":{"tags":["service-controller"],"operationId":"deleteService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/organisations/IsDeleted/{id}":{"delete":{"tags":["organisation-controller"],"operationId":"deleteOrganisation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/offices/deleteOffice/{id}":{"delete":{"tags":["office-controller"],"operationId":"deleteOffice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Feedback":{"type":"object","properties":{"feedbackId":{"type":"integer","format":"int64"},"note":{"type":"number","format":"double"},"comment":{"type":"string"},"organisation":{"$ref":"#/components/schemas/Organization"},"user":{"$ref":"#/components/schemas/User"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"read":{"type":"boolean"}}},"Organization":{"type":"object","properties":{"organizationId":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"type":"string"},"phone_number":{"type":"string"},"email":{"type":"string"},"image":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"integer","format":"int32"},"averageRate":{"type":"number","format":"float"},"is_deleted":{"type":"boolean"}}},"User":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"first_name":{"type":"string"},"last_name":{"type":"string"},"full_name":{"type":"string"},"birth_date":{"type":"string","format":"date"},"image":{"type":"string"},"address":{"type":"string"},"password":{"type":"string"},"reset_pwd_code":{"type":"string"},"reset_pwd_date":{"type":"string","format":"date-time"},"phone_number":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["CLIENT","ADMIN","STAFF"]},"barCode":{"type":"string"},"enabled":{"type":"boolean"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"},"organization":{"$ref":"#/components/schemas/Organization"},"is_locked":{"type":"boolean"},"is_deleted":{"type":"boolean"}}},"Services":{"type":"object","properties":{"serviceId":{"type":"integer","format":"int64"},"serviceName":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"integer","format":"int32"}}},"Office":{"type":"object","properties":{"officeId":{"type":"integer","format":"int64"},"name":{"type":"string"},"location":{"type":"string"},"phoneNumber":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"integer","format":"int32"}}},"OrganizationEntityCountsDTO":{"type":"object","properties":{"organizationId":{"type":"integer","format":"int64"},"organizationName":{"type":"string"},"officeCount":{"type":"integer","format":"int32"},"feedbackCount":{"type":"integer","format":"int32"},"userCount":{"type":"integer","format":"int32"}}},"FeedBackDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"note":{"type":"number","format":"double"},"comment":{"type":"string"},"read":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}